+++ carboculator-frontend/postcss.config.js
@@ -1,6 +1,7 @@
+// postcss.config.js
+import tailwindcss from '@tailwindcss/postcss'
+import autoprefixer from 'autoprefixer'
+
 export default {
-  plugins: {
-    tailwindcss: {},
-    autoprefixer: {},
-  },
+  plugins: [tailwindcss, autoprefixer],
 }
