+++ carboculator-frontend/src/App.jsx
@@ -5,6 +5,9 @@
   import LoginPage from "./Components/AccountSettings/LoginPage"; 
 import Visualization from "./Components/DashBoard/Visualization/Visualization";
 import CarbonCredits from "./Components/DashBoard/CarbonCredits";
+import RegistrationPage1 from "./Components/RegistrationPage/RegistrationPage1";
+import RegistrationPage2 from "./Components/RegistrationPage/RegistrationPage2";
+import CreateAccount from "./Components/RegistrationPage/CreateAccount";
 
   
   const App = () => {
@@ -21,7 +24,10 @@ import CarbonCredits from "./Components/DashBoard/CarbonCredits";
       {/* <SideBar/> */}
      {/* <Visualization />
     */}
-    <CarbonCredits />
+    {/* <CarbonCredits /> */}
+      {/* <RegistrationPage1 /> */}
+      <RegistrationPage2 />
+      {/* <CreateAccount /> */}
 
       </>
