+++ carboculator-frontend/src/index.css
@@ -1,0 +1,4 @@
+/* src/index.css */
+
+@import "tailwindcss"
+
